wip: refactor resolve
